extract dashboard (un)install from theme to layout
