+++ src/ploneintranet/layout/profiles/default/actions.xml
