+++ src/ploneintranet/layout/setuphandlers.py
@@ -1,18 +1,21 @@
 # -*- coding: utf-8 -*-
+from plone import api
+
 import logging
 log = logging.getLogger(__name__)
 
 
-def isNotPolicy(context):
-    return context.readDataFile("ploneintranet.theme.txt") is None
+def setupVarious(context):
+    if context.readDataFile("ploneintranet.layout_default.txt") is None:
+        return
+
+    configureFrontPage(context)
 
 
 def configureFrontPage(context):
     """ Delete the "Welcome to Plone" page and
         set the dashboard as default view
     """
-    if isNotPolicy(context):
-        return
     site = context.getSite()
     if "front-page" in site.objectIds():
         site.manage_delObjects(['front-page'])
@@ -23,3 +26,10 @@ def configureFrontPage(context):
         site.manage_addProperty(id='layout',
                                 value='dashboard.html',
                                 type='string')
+
+
+def uninstall(context):
+    if context.readDataFile('ploneintranet.layout_uninstall.txt') is None:
+        return
+    portal = api.portal.get()
+    portal.setLayout('listing_view')
