+++ src/ploneintranet/theme/configure.zcml
@@ -16,7 +16,6 @@
 
     <include package=".browser" />
 
-    <!-- Register the installation GenericSetup extension profile -->
     <genericsetup:registerProfile
         name="default"
         title="Plone Intranet: -- [Theme]"
@@ -33,29 +32,18 @@
         provides="Products.GenericSetup.interfaces.EXTENSION"
         />
 
-    <genericsetup:importStep
-        name="ploneintranet.theme.configureFrontPage"
-        title="Delete default frontpage and configure dashboard as default"
-        description=""
-        handler="ploneintranet.theme.setuphandlers.configureFrontPage">
-    </genericsetup:importStep>
-
-    <genericsetup:importStep
-        name="ploneintranet.theme.uninstall"
-        title="Plone Intranet Theme: uninstall steps"
-        description="Various uninstall steps that are not handled by GS import/export handlers."
-        handler="ploneintranet.theme.uninstall.uninstall" />
-
     <plone:static
         directory="static"
         type="theme"
         name="ploneintranet.theme"
         />
-    <include package="z3c.jbot" file="meta.zcml" />
+
     <interface name="ploneintranet.theme"
         interface="ploneintranet.theme.interfaces.IThemeSpecific"
         type="zope.publisher.interfaces.browser.IBrowserSkinType"
         />
+
+    <include package="z3c.jbot" file="meta.zcml" />
     <browser:jbot directory="static/overrides"
         layer=".interfaces.IThemeSpecific" />
